+++ astroNN/models/base_bayesian_cnn.py
@@ -19,7 +19,7 @@ from astroNN.nn.metrics import categorical_accuracy, binary_accuracy
 from astroNN.nn.numpy import sigmoid
 from astroNN.nn.utilities import Normalizer
 from astroNN.nn.utilities.generator import GeneratorMaster
-from astroNN.shared.warnings import deprecated
+from astroNN.shared.warnings import deprecated, deprecated_copy_signature
 from astroNN.shared.nn_tools import gpu_availability
 from astroNN.shared.dict_tools import dict_np_to_dict_list, list_to_dict
 
@@ -336,7 +336,7 @@ class BayesianCNNBase(NeuralNetMaster, ABC):
 
         return {m.name: m.result() for m in self.keras_model.metrics}
 
-    def train(self, input_data, labels, inputs_err=None, labels_err=None, sample_weights=None):
+    def fit(self, input_data, labels, inputs_err=None, labels_err=None, sample_weights=None):
         """
         Train a Bayesian neural network
 
@@ -400,7 +400,7 @@ class BayesianCNNBase(NeuralNetMaster, ABC):
 
         return None
 
-    def train_on_batch(self, input_data, labels, inputs_err=None, labels_err=None, sample_weights=None):
+    def fit_on_batch(self, input_data, labels, inputs_err=None, labels_err=None, sample_weights=None):
         """
         Train a Bayesian neural network by running a single gradient update on all of your data, suitable for fine-tuning
 
@@ -509,7 +509,7 @@ class BayesianCNNBase(NeuralNetMaster, ABC):
         with open(self.fullfilepath + '/astroNN_model_parameter.json', 'w') as f:
             json.dump(data, f, indent=4, sort_keys=True)
 
-    def test(self, input_data, inputs_err=None):
+    def predict(self, input_data, inputs_err=None):
         """
         Test model, High performance version designed for fast variational inference on GPU
 
@@ -726,3 +726,16 @@ class BayesianCNNBase(NeuralNetMaster, ABC):
         print(f'Completed Evaluation, {(time.time() - start_time):.{2}f}s elapsed')
 
         return list_to_dict(funcname, scores)
+    
+
+    @deprecated_copy_signature(fit)
+    def train(self, *args, **kwargs):
+        return self.fit(*args, **kwargs)
+
+    @deprecated_copy_signature(fit_on_batch)
+    def train_on_batch(self, *args, **kwargs):
+        return self.fit_on_batch(*args, **kwargs)
+    
+    @deprecated_copy_signature(predict)
+    def test(self, *args, **kwargs):
+        return self.predict(*args, **kwargs)