+++ astroNN/models/base_vae.py
@@ -15,6 +15,7 @@ from astroNN.nn.losses import mean_squared_error, mean_error, mean_absolute_erro
 from astroNN.nn.utilities import Normalizer
 from astroNN.nn.utilities.generator import GeneratorMaster
 from astroNN.shared.dict_tools import dict_np_to_dict_list, list_to_dict
+from astroNN.shared.warnings import deprecated, deprecated_copy_signature
 from sklearn.model_selection import train_test_split
 
 regularizers = tfk.regularizers
@@ -252,7 +253,7 @@ class ConvVAEBase(NeuralNetMaster, ABC):
 
         return input_data, input_recon_target
 
-    def train(self, input_data, input_recon_target, sample_weights=None):
+    def fit(self, input_data, input_recon_target, sample_weights=None):
         """
         Train a Convolutional Autoencoder
 
@@ -300,7 +301,7 @@ class ConvVAEBase(NeuralNetMaster, ABC):
 
         return None
 
-    def train_on_batch(self, input_data, input_recon_target, sample_weights=None):
+    def fit_on_batch(self, input_data, input_recon_target, sample_weights=None):
         """
         Train a AutoEncoder by running a single gradient update on all of your data, suitable for fine-tuning
 
@@ -386,7 +387,7 @@ class ConvVAEBase(NeuralNetMaster, ABC):
         with open(self.fullfilepath + '/astroNN_model_parameter.json', 'w') as f:
             json.dump(data, f, indent=4, sort_keys=True)
 
-    def test(self, input_data):
+    def predict(self, input_data):
         """
         Use the neural network to do inference and get reconstructed data
 
@@ -459,7 +460,7 @@ class ConvVAEBase(NeuralNetMaster, ABC):
 
         return predictions
 
-    def test_encoder(self, input_data):
+    def predict_encoder(self, input_data):
         """
         Use the neural network to do inference and get the hidden layer encoding/representation
 
@@ -573,3 +574,19 @@ class ConvVAEBase(NeuralNetMaster, ABC):
         print(f'Completed Evaluation, {(time.time() - start_time):.{2}f}s elapsed')
 
         return list_to_dict(funcname, scores)
+
+    @deprecated_copy_signature(fit)
+    def train(self, *args, **kwargs):
+        return self.fit(*args, **kwargs)
+
+    @deprecated_copy_signature(fit_on_batch)
+    def train_on_batch(self, *args, **kwargs):
+        return self.fit_on_batch(*args, **kwargs)
+    
+    @deprecated_copy_signature(predict)
+    def test(self, *args, **kwargs):
+        return self.predict(*args, **kwargs)
+    
+    @deprecated_copy_signature(predict)
+    def test_encoder(self, *args, **kwargs):
+        return self.predict_encoder(*args, **kwargs)